Add(halam/container-presenter):applied container presenter pattern #43
