+++ moae-time/src/pages/home/HomePageContainer.jsx
@@ -1,0 +1,17 @@
+import React, { useState } from 'react';
+import HomePagePresenter from './HomePagePresenter';
+
+function HomePage() {
+  const [modalVisible, setModalVisible] = useState(false);
+  const handleModalState = () => {
+    setModalVisible(!modalVisible);
+  };
+  return (
+    <HomePagePresenter
+      modalVisible={modalVisible}
+      handleModalState={handleModalState}
+    />
+  );
+}
+
+export default HomePage;
