+++ moae-time/src/pages/index.js
@@ -1,5 +1,5 @@
-import HomePage from "./home/HomePage";
-import BoardPage from "./board/BoardPage";
-import MyPage from "./mypage/MyPageContainer";
+import HomePage from './home/HomePageContainer';
+import BoardPage from './board/BoardPage';
+import MyPage from './mypage/MyPageContainer';
 
 export { HomePage, BoardPage, MyPage };
